feat(editor): update event attributes and add EventEditor
The update includes changes in the "payload" and "debug" attributes from private to public in the Event class. This provides greater control when initiating events from the inspector. Also, an "EventEditor" class was added to extend the functionality of custom event editors. This includes a button to invoke the defined event directly from the Unity Editor Inspector. The root namespace in the "asmdef" file was also adjusted accordingly.

Signed-off-by: Alex Ruiz <68024691+aruizrab@users.noreply.github.com>