chore(events): refactor documentation and adjust namespace in tests
The payload type documentation in `Event.cs` is updated for clarity, from `T` to `<typeparamref name="T" />`. Also, modified the structure of an example sentence for better readability. In `EventTest.cs` and `com.github.aruizrab.sodd-unity-framework.Tests.asmdef`, changes are made to the namespace of the test suite to better fit the standard convention.

Signed-off-by: Alex Ruiz <68024691+aruizrab@users.noreply.github.com>