feat(variables): add abstract Variable class for scriptable variables
The commit introduces the Variable abstract class for the Unity Framework of the SODD package. This generic class will serve as the base for all scriptable variable types, allowing data containers that can be easily manipulated within Unity and referenced across scripts. Additionally, unit tests have been added to ensure the correct operation and usage of the Variable class.

Signed-off-by: Alex Ruiz <a.ruiz.rabasseda@gmail.com>