feat(input): add abstract InputActionHandler class for scriptable action handlers
Implemented a new class `InputActionHandler` under Runtime/Input/ActionHandlers for handling various types of input actions. This class is responsible for managing input actions and triggering appropriate scriptable events in response. Moreover, added `Unity.InputSystem` as a reference in the assembly definition file and as a dependency in the package.json file to facilitate the usage of Unity's input system within the framework.

BREAKING CHANGE: `Unity.InputSystem` is now a package dependency

Signed-off-by: Alex Ruiz <a.ruiz.rabasseda@gmail.com>